refactor(jobs): split up JobsTab component
this commit splits up the old JobsTab component into a set of 'dumb' UI components (in JobsTab.js) and a container component (JobsTabContainer.js).

This is the first step of multiple steps to refactor the Jobs Overview Page.

Closes DCOS-37750
