+++ plugins/jobs/src/js/components/JobSearchFilter.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars */
 import PropTypes from "prop-types";
-
 import React from "react";
-/* eslint-enable no-unused-vars */
-import FilterInputText from "./FilterInputText";
+
+// everything that is references as #ALIAS/something has to be refactored once our DI system is in place
+import FilterInputText from "#SRC/js/components/FilterInputText";
 
 class JobSearchFilter extends React.Component {
   render() {
