+++ plugins/jobs/src/js/components/JobsTable.js
@@ -5,13 +5,15 @@ import PropTypes from "prop-types";
 import React from "react";
 import { Table, Tooltip } from "reactjs-components";
 
-import Icon from "../../components/Icon";
-import JobStates from "../../constants/JobStates";
-import JobStatus from "../../constants/JobStatus";
-import JobTableHeaderLabels from "../../constants/JobTableHeaderLables";
-import ResourceTableUtil from "../../utils/ResourceTableUtil";
-import TableUtil from "../../utils/TableUtil";
-import Tree from "../../structs/Tree";
+// everything that is references as #ALIAS/something has to be refactored once our DI system is in place
+import Icon from "#SRC/js/components/Icon";
+import ResourceTableUtil from "#SRC/js/utils/ResourceTableUtil";
+import TableUtil from "#SRC/js/utils/TableUtil";
+import Tree from "#SRC/js/structs/Tree";
+
+import JobStates from "../constants/JobStates";
+import JobStatus from "../constants/JobStatus";
+import JobTableHeaderLabels from "../constants/JobTableHeaderLabels";
 
 const METHODS_TO_BIND = ["renderHeadline", "jobSortFunction"];
