+++ plugins/jobs/src/js/constants/JobFilterTypes.js
