+++ plugins/jobs/src/js/constants/JobStates.ts
