+++ plugins/jobs/src/js/constants/JobStatus.ts
