+++ src/js/pages/jobs/JobDetailPage.js
@@ -6,11 +6,11 @@ import React from "react";
 import { routerShape } from "react-router";
 import { Confirm } from "reactjs-components";
 import mixin from "reactjs-mixin";
-import TaskStates from "../../../../plugins/services/src/js/constants/TaskStates";
+import TaskStates from "#PLUGINS/services/src/js/constants/TaskStates";
+import JobsBreadcrumbs from "#PLUGINS/jobs/src/js/components/JobsBreadcrumbs";
 import Icon from "../../components/Icon";
 import Page from "../../components/Page";
 import TimeAgo from "../../components/TimeAgo";
-import JobsBreadcrumbs from "../../components/breadcrumbs/JobsBreadcrumbs";
 import JobFormModal from "../../components/modals/JobFormModal";
 import UserActions from "../../constants/UserActions";
 import TabsMixin from "../../mixins/TabsMixin";
